Calculate answers in `long long` to avoid overflow

+++ day-32-cf-706-c.cpp
@@ -30,18 +30,18 @@ int main()
      * min_original[i] stores the minimum number of reversals required to achieve non-descending order of first (i + 1)
      * words with `i`th word not reversed (`INT32_MAX` if not achievable)
      */
-    std::vector<int> min_original((size_t)num_words);
+    std::vector<long long> min_original((size_t)num_words);
 
     /**
      * Same, but with the `i`th word reversed
      */
-    std::vector<int> min_reversed((size_t)num_words);
+    std::vector<long long> min_reversed((size_t)num_words);
 
     min_original[0] = 0;
     min_reversed[0] = reversal_costs[0];
 
     for(int i = 1; i < num_words; i++) {
-        min_original[i] = INT32_MAX;
+        min_original[i] = INT64_MAX;
 
         if(words[i] >= words[i - 1]) {
             min_original[i] = std::min(min_original[i], min_original[i - 1]);
@@ -51,20 +51,20 @@ int main()
             min_original[i] = std::min(min_original[i], min_reversed[i - 1]);
         }
 
-        min_reversed[i] = INT32_MAX;
+        min_reversed[i] = INT64_MAX;
 
-        if(reversed_words[i] >= words[i - 1] && min_original[i - 1] != INT32_MAX) {
+        if(reversed_words[i] >= words[i - 1] && min_original[i - 1] != INT64_MAX) {
             min_reversed[i] = std::min(min_reversed[i], min_original[i - 1] + reversal_costs[i]);
         }
 
-        if(reversed_words[i] >= reversed_words[i - 1] && min_reversed[i - 1] != INT32_MAX) {
+        if(reversed_words[i] >= reversed_words[i - 1] && min_reversed[i - 1] != INT64_MAX) {
             min_reversed[i] = std::min(min_reversed[i], min_reversed[i - 1] + reversal_costs[i]);
         }
     }
 
-    int result = std::min(min_original[num_words - 1], min_reversed[num_words - 1]);
+    long long result = std::min(min_original[num_words - 1], min_reversed[num_words - 1]);
 
-    if(result != INT32_MAX) {
+    if(result != INT64_MAX) {
         std::cout << result << std::endl;
     }
     else {
